Change scale up strategy to use scale up steps at the connector level
The strategy to compute a new desired state, depends on the accurate knowledge of the
current desired instance group size, which can be changed by multiple sources,
and cannot be ever known with absolute certenity. It is safer to scale up by deltas,
with the risk of scaling too much, which is better than accidental scale down possible
when desired size is changed directly.