Task migration fails on v3 task when it is already finished (#140)
A finished task always stays in Fenzo queue for some time. In V2 engine as we use mutable
objects it was not an issue. In V3 engine we need to ask for the latest version of the
job/task which may no longer be in the active data set.

In V3 engine, as everything runs on a reconciler loop, there is some delay between the time
the task state changes to 'Finished', and the time the task is removed from the Fenzo queue.
We could optimize it, and remove from Fenzo much earlier, and thus reduce the delay
from >100ms to single Fenzo loop evaluation time. But this delay we cannot eliminate it
completely. Instead we should change the task migrator implementation to handle this fact directly.