federation: findJobs with cursor pagination
Fan out and query all cells with a cursor. The current estimation for
pageNumber is (very) inaccurate, and will need more information from
each cell to be refined. That will come in a future patch.

This also changes the semantics of Pagination.currentPage when cursors
are used. It is set to be the page that includes the requested cursor,
instead of the requested pageNumber since that is ignored when a cursor
is used.