cursor pagination: cursor position and federated pageNumber estimation
This depends on titus-api-definitions#29 being merged.

Set the cursor position on each cell, and use it to provide a sense of
progress for clients using cursor based pagination, and more accurately
estimate the current aggregated pageNumber when cursors are being used
with federated cells.

The new tests also found bugs with the previous cursor-based pagination,
fixed here:

1) fix the min cursorIndexOf, which was causing items to be missed when
cursors point to items in the beginning of each Cell's list

2) always return a cursor when there are items, if the current page is
empty, the returned cursor (and cursor position) will still point to the
last existing item, instead of being empty. This is more correct, and
necessary for federation to determine the aggregate totals when fetching
the last pages from each cell.