federation: findTasks, extract page aggregation and cell/grpc logic
First pass at extracting some common federated pagination logic, and
grpc call to cells. This is an moving target and I expect more
iterations as things evolve.

This also adds a federated findTasks implementation using the new
common extracted logic. Tests and the endpoints (grpc and rest) for it
will come in a followup patch.