federation: response merging logic for fan out and point queries
Implement the remaining endpoints and add some common logic to merge
responses and errors back.

observeJob(id) is the only endpoint missing, and will be implemented in
a future patch.

AggregatingAutoscalingService was not modified. Changes to make it use
the new response merging logic will come in a future patch as well.