let the federation proxy set titus.stack for jobs and tasks
This makes it so jobs and tasks are always persisted with `titus.stack`
being the federated stack name. The caveat is that jobs created not
through the federation proxy (including legacy V2 API calls that don't
go through federation) will not have the attribute persisted with them.
With it, the persisted state should more accurately reflect what jobs
and tasks were created in a federated context, and what is their
federated stack name.

All jobs and tasks queried through a federation proxy continue to have
`titus.stack` automatically decorated into them, regardless if the
records from each cell have it or not.