Rewrite fromWithDelay() to avoid an unsound cast
Observable.timer() has type Observable<Long>, and ignoreElements() also
returns Observable<Long>. The cast to Observable<T> is thus unsound,
although it appears to not make a difference because ignoreElements()
discards the Long value generated by the timer source.

The rewritten version also does not use recursion and avoids calling
size() on the chunks List, which might be expensive depending on the
List implementation.