Add job validator to gateway
There are types of validation for Jobs which extend beyond the currently
available syntactic sanitization/validation approach.  Here we provide a
method for providing arbitrary validation of Jobs upon creation at the
gateway layer.  If a job fails validation the Titus master component
will never attempt to launch the job and clients will be notified of
failure.