allow after_* hooks to run on travis
calling `exit` on failures was making the CI script on travis from
running the after_* hooks, and dumping logs on failures.

By running the curl tests in a subshell, we prevent the CI script
itself from exiting.

This also reduces the time we wait for docker containers to be
UP.