prevent unavailable Mesos from crashing the JVM
DefaultSchedulingService will System.exit on any exceptions coming
from the Fenzo scheduling loop. In case Mesos is unavailable and
offers can't be rejected, this can cause Titus master to crash.

Since Mesos is unavailable, we ignore the fact that offers can't
be declined for now, and rely on re-registration (and lease
expirations) to later fix offer states when Mesos comes back up.