limit attempts for opportunistic scheduling
When opportunistic scheduling fails, we reduce the number of allocated
opportunistic CPUs for the next iteration. When it reaches zero for the
first time, stop opportunistic scheduling altogether (set
opportunisticCpus to 0 permanently), instead of retrying from the
beginning with all requested CPUs being opportunistic.

This prevents any scheduling constraint failures not related to
opportunistic scheduling from stopping tasks to be scheduled completely.