limit threads used by ObserveJobs and add instrumentation
* use a single thread pool for all gRPC tasks, instead of a separate
thread pool per channel (default with gRPC)
* limit the number of threads used for server streaming (ObserveJobs)
tasks, this will cause too many calls to sit on a queue when snapshots
can't be generated fast enough (requests are queued when we are
overloaded)
* metrics for thread pool sizes, busy thread counts, and task counts on
gRPC thread pools

I did some load testing and an empty Titus stack with these changes can
sustain ~300 ObserveJobs calls/sec with little thread pool usage. The
only remaining problem is with unsubscribe() (when gRPC streams are
canceled), which is blocking inside the RxJava stack, and putting a lot
of pressure on gRPC thread pools and the Rx Computation Scheduler.
Canceling 100k ObserveJobs streams consumed ~4k threads for a few seconds
and caused the hiccup sensor to complain about threads blocked in the Rx
Computation Scheduler.