Read archived data from active job/task tables if not found in the archive tables (#868)
* Read archived data from active job/task tables if not found in the archive tables

It is possible that finished jobs/tasks are not moved after they are finished to
the archive tables. This may happen during failover which happens
immediately after a job completes, but is not archived yet. To read such
abandoned items (which are ignored when loading active data into memory), we
have to make a fallback query in case the primary archive table does not contain
an entry.

* Code review updates