ExclusiveHost has two different failure modes
A failure for a task that has the hard constraint is different from
other tasks failing because a machine is already running a task with
ExclusiveHost. Both failures are classified with different priority to
avoid prematurely triggering the ClusterAgentAutoscaler when tasks
that don't have the hard constraint are failing due to other transient
scheduling failures.