Add a sanitizer for service mesh image attribute.
We expect the gateway to replace tags with digest when launching a
job with service mesh container specified.

This allows the service networking team to have customers launch a
service job with a release build, while making sure that all tasks
launched for the job use the same version of proxyd.