opportunistic CPUs: faster fallback from low supply
With jobs elligible for opportunistic scheduling, reduce the number of
allocated opportunistic CPUs by half on each failed scheduling
iteration, accelerating fallback to regular scheduling
(opportunisticCpus = 0) when there is not enough surplus of opportunistic
CPUs available.

This trades off optimal usage of opportunistic CPUs for lower launch
latencies, as opportunsitic tasks will sit on the queue for a shorter
time waiting for opportunistic CPUs to show up.

It also introduce some bias in the amounts of opportunistic CPUs that
are allocated to each task (e.g.: prime numbers, and number not
divisible by 2 may be less likely to be used), but the tradeoff in queue
times and reliability (during a large spike of opportunistic tasks and
short supply) is worth for now.