metrics: tag task update counters with applicationName
This is helpful in determining what applicationName is responsible for
big spikes (surges) in task launches.

While in theory applicationName has an unbound cardinality, and is a
user-provided value that could cause issues with high cardinality
metrics, in practice we already have other counter metrics in the system
using it as a tag, and the cardinality in live production systems tends
to be manageable.