Added new NetworkConfiguration API
Pairs with
https://github.com/Netflix/titus-api-definitions/pull/130

This represents a new place to put job-level networking
configuration (instead of container resources).
Starting with a new enum for networking mode.

My intent here is only to pass this through into the pod
for the backend to consume, no other logic in the control-plane
should be needed here.