Only set the ipv4 container attribute if we have one
We are moving to a new world where IPv6 is the main thing,
and IPv4 is extra.

This change makes the ipv4/ipv6 task attribute setter "NetworkMode"-aware,
and only sets the IPv4 when it makes sense.

Only a small number of experimental tasks are using this mode,
so we need to use those as canaries to see what else will break
when we stop publishing a v4 address in this way.