Implemented extraContainers
This changed takes in the new proto for ExtraContainers and appends
those containers to the pod (v0 and v1 schemas).

This adds a new place for adding test to the V1 schema pod builder
as well.

These new extraContainers are constructed pretty much as-is from
proto to V1Container, minimal processing is done.