Take incoming platform Sidecars and make annotations for them (#1115)
* WIP: Take incoming platform Sidecars and make annotations for them

This new feature takes the new PlatformSidecar objects from the
Titus API proto definition and populates the appropriate
annotations on the pod. No additional processing is required by
the control-plane, because we want to stay pretty close to
the "the pod spec is the api" vision.

Two big things we need to iron out:

1. Annotation format. It would be nice to get some good conventions,
and hopefully not change them over time. Luckily in the webhook
we can parse multiple annotations and merge them together as we change.

2. This whole `Struct` stuff. The reasoning behind using a protobuf
[Struct](https://developers.google.com/protocol-buffers/docs/reference/google.protobuf#google.protobuf.Struct)
field was to preserve the structured data that would come in
from the API, and not treat it like a "string" of escaped json.

I think this is still probably a good idea, and only stringify
it out at the last second sounds also like a good idea?
But I've never done something like this in java, so it needs
extra scrutiny. Also I don't know if it is kosher for the actual
POJO to "just" be a struct, but it sure simplifies things!

* Correctly serialize platform sidecars into json for cassandra

* Small cleanup

* Use json strings as the fundamental unit of platform sidecar arguments

* Use Either<> to represent platform sidecar arguments or errors

* Used TitusServiceException when bad arguments are encounterd with platform sidecars

* Minor fixes after rebase on master