Fixed v1 pod spec support on efs mounts
This thing is a little tricky, because on Titus
we have an "efs mount", but on the pod that needs to
be correctly decomposed into volume/volumemounts.

Previously it was not 100% correct and the titus
functional tests were failing on all efs tests.

With this change, they pass.

Additionally, this particular way of naming things
should future-proof this for when one has multiple
efs mounts with weird different mount points
on different containers (multi-container workloads).