Updated the titus-kube-scheduler-reserved-binpacking string
I find it a bit confusing that the default for this string *isn't*
`titus-kube-scheduler-reserved-binpacking`. We have properties set
in many places (not all?) that "overrides" it to be correct.

I think it might be easier to understand if it is correct here,
and we ony set a property to override the behavior.

If shipit, after this change is deployed, I'll delete the properties.