Sanitize v1 pod efs volumes that end in a '/'
All volume names in k8s need to end on a normal character.

Before, a trailling '/' (or the default relative path if unset!)
would leave the volume name with a trailing '-'.

With this change, we always add `-vol` at the end of the volume name,
regardless of what wacky mount points we have.