Added support for filtering based on platformSidecar channel
This is part2 of allowing platform sidecar (PS) authors to be
able to find jobs of interest to them.

Specifically, this will allow them to find jobs that are using
a particular channel, so those jobs can be bounced or hunted
down when a particular channel needs to be deprecated.

*Or* when that channel is upgraded, jobs/tasks using that channel
can be targeted for relocation, to force them to get the latest
version.

----

I didn't feel the need to go "all out" and write lots of integration tests
to verify this new code's interaction with "platformSidecar=", I think it
can be safely assumed that, like any other query filter, this is an "And"
filter, and will be applied *together* with a sidecar name to narrow the scope.

While technically possible for one to just query based on a channel name,
seems like it would be a little silly.