Convert EFS IDs to NFS Hostnames for v1 pod specs
Previously, in v0 pod specs the Titus executor would
do a conversion at the last minute, extracting the
efsid from the protobuf.

In v1 pod specs, I think it makes more sense for the
nfs:server field to be "true" and actually be an NFS
server, which means implement this conversion in the
control plane.