Make the name of the 'main' container be 'main' in V1 schema pods (#1162)
I want to make this name change before we have too many things
keying off of container names in the newer multi-container world.

Remember, we already have a unique identifier for the pod, the pod ID
is the task uuid, that is great.

The container name though, I think it would be better to have a consistent
string we can key off of: "main". This simplifies a lot of future code
that has to detect "which" container they are on, or making a volumeMount
that needs to reference "which" container to attach.

For all things that ever need to ask "what is the name of the main container"
or "am I running on the main container", they will not need to *also* look
up the taskid var and compare.