v1 pod schema: Handle multiple complex efs volumes
Previously, we did not handle the case where we had
multiple volumes with a mix of RO/RW on the same source.

This change makes it so we always put the burden of RW/RO
on the volume *mount* (not the volume), and re-use the
volume itself (always RW, doesn't really matter too much).