Remove unused EBS PV code
We had this code in previously for kubelet + the
AWS EBS CSI Driver.

Now that we are no longer using that driver, we no longer
need the complexity of PV / PVCs at all (pretty sure),
and all we need is a plain volume + volumeMount.

Also with this change, EBS volumes will be enabled in all
regions, and we'll no longer need a feature toggle.