Use EBS annotations for v0, and vol/volmount for v1 pods
Previously we were always creating annotations *and* vol/volmount
for EBS on v0 *and* v1 pods. I found this confusing, because
annotations are used on v0 and vols are used on v1.

I didn't want to perpetuate the annotations (wasting space)
for no reason on v1 pods.

This change makes it more clear that on v0 we use annotations
only, and on v1 we use pod constructs only.

This also extends the tests to *actually* create a pod, where
as before we were only calling a private function, which of course
always returns v1 volumes.