Use awsElasticBlockStore instead of PVS pods
This change makes it so we always use the awsElasticBlockStore
volumeSource on a volume, instead of using PV/PVCs.

This does leave a lot of leftover unused code that I will clean up on
the next PR, only after verifying there are no unintended side-effects
from configuring pods this way.

On v0 pods, nothing should change, because we still just use the
annotations as the source of truth.

For v1 pods, tests should go green (they fail currently), because the
titus-executor expects awsElasticBlockStore to be filled in, but is
currently not.