Relax Job validation on the HighScale network mode
The point of the HighScale network mode is to set
security groups and subnets on the backend, chosen at
task start time by our infrastructure, *not* baked into the job.

to that end, we must relax the validation when in this mode.

Note that I did not add *additional* validation that SG/subnets are
*not* set when in this mode. I will be doing that in a subsequent
PR in the new titus-federation codebase (I think).