Make v0 pods use the "main" container as the container name
This is the same as v1 pods, but bringing this feature to v0 pods
because I don't really see us getting to v1 pods any time soon.

But, platform sidecar developers need this feature, because we need to
be able to reference the main container "by name", and it is way easier
when that is a constant.