Use the existing containerEnv factory with V1 pods
I want v1 pods to use the existing business logic as v0 pods when it
comes to generating env variables (at least, as much as possible).

This change allows V1 pods to get the existing env logic, which includes
setting private NETFLIX env variables.

To do this, I had to extend the API of the buildContainerEnv to export
the pair thing, even if in v0 it doesn't do anything with the left hand
side of the pair.