Added image data and platform sidecar data to containerStatus
Now that we have containerStatus plumbed all the way through, this is
the next pass which adds more data like image stuff and platform sidecar
stuff.

The reason this is difficult is because with platform sidecars, we need
to look at annotations to verify what the exact tag/digest/name that an
image comes from. The stock k8s v1 container status has no room for this
of course. And it doesn't have room for tag/digest, which we really
want.

This adds code for parsing and extracting annotations to be able to get
back the exact container image and (optionally) the platform sidecar
that added it.

I consider this code to be "best effort", as this is a new feature and
there are a number of ways that the data can simply not show up.

Once this data is in our API, I can make it visible to the Titus UI and
spinnaker to make it clear to our users exactly what sidecars are
running, what version, and why (like what platform sidecar mutator added
it).