RFC: Added more swagger annotations, hid others
I was looking at our API docs and wondered what it might take to get our
swagger page into a position where it could replace our (outdated)
markdown docs.

I don't think it is too bad. In fact I think it is good that the params
and what they do are so closely tied to the docs like this.

I still can't hide `authorities[0].authority` and 6 other params that
are crowding out the page, but maybe on the next pass?