Use a special EIP task context attribute instead of the primary IP
I incorrectly added code in https://github.com/Netflix/titus-control-plane/pull/1220/
that made the EIP of a task be the "primary" ip.

This broke LB registration.

I liked the idea of getting the EIP context data back out, but it should
be the primary ip. Instead this adds a new piece of context data so that
one can still see what the EIP was in the UI and other tools, but the
primary ip will remain the private ipv4 address.